feat(search-results): moving results page into a specific folder
